Remove AttributeList from Device Layer (#41546)
* Remove AttributeList from Device Layer

It was only used in a single place and probably didn't belong in the device layer
in the first place.  Use a Span instead for DeviceInforProvider::SetUserLabelList.

Also simplify DeviceInforProvider::ClearUserLabelList().

* Add missing const to Span type

* Move max size constants into the header and add more tests

* Allow the global DeviceInfoProvider to be reset to null

This is primarily useful for testing. Also remove the comment stating that
callers should assume the provider is not null; it starts out as null, and the
cluster code interacting with it already handles null gracefully.

---------

Co-authored-by: Andrei Litvin <andy314@gmail.com>